flask/requirements.txt
flask==3.0.2
flask-cors==3.0.10
requests==2.31.0
numpy==1.24.3
pillow==10.2.0
tensorflow==2.15.0
keras==3.9.1
gunicorn==21.2.0
gevent==23.9.1
python-dotenv==1.0.0
prometheus-client==0.18.0